+++ ecommers/lib/ui/widgets/right_menu_bar/models/right_menu_item_model.dart
@@ -1,4 +1,10 @@
+import 'package:flutter/widgets.dart';
+
 class RightMenuItemModel {
-   final String title;
-   RightMenuItemModel({this.title});
+  final String title;
+  RightMenuItemModel({this.title});
+
+  Widget create() {
+    return const Text('');
+  }
 }
