+++ ecommers/lib/ui/widgets/right_menu_bar/right_menu_item.dart
@@ -4,10 +4,7 @@ import 'package:flutter_svg/svg.dart';
 import 'package:ecommers/ui/decorations/index.dart';
 import 'package:ecommers/ui/decorations/dimens/index.dart';
 
-import 'models/colors_model.dart';
-import 'models/price_model.dart';
 import 'models/right_menu_item_model.dart';
-import 'models/subtitle_model.dart';
 
 @immutable
 class RightMenuItem extends StatelessWidget {
@@ -40,57 +37,7 @@ class RightMenuItem extends StatelessWidget {
 
 @override
 Widget _subTitleWidget(BuildContext context, RightMenuItemModel itemModel) {
-  if (itemModel is ColorsModel) {
-    final listColors = itemModel;
-
-    return Stack(
-        children: listColors.colors
-            .map((color) =>
-                circleBuilder(color, listColors.colors.indexOf(color)))
-            .toList());
-  }
-  if (itemModel is SubTitleModel) {
-    final subTitleModel = itemModel;
-    return Text(
-      subTitleModel.subTitle,
-      style: Theme.of(context)
-          .textTheme
-          .subtitle2
-          .copyWith(fontWeight: FontWeight.w200),
-    );
-  }
-  if (itemModel is PriceModel) {
-    final priceModel = itemModel;
-    final _minPrice = priceModel.minPrice;
-    final _maxPrice = priceModel.maxPrice;
-
-    return Text(
-      '\$ $_minPrice - \$ $_maxPrice',
-      style: Theme.of(context)
-          .textTheme
-          .subtitle2
-          .copyWith(fontWeight: FontWeight.w200),
-    );
-  }
-  return const Text('null');
+  return itemModel.create();
 }
 
-Widget circleBuilder(String color, int index) {
-  return Row(children: <Widget>[
-    Container(
-        width: Insets.x5,
-        height: Insets.x5,
-        margin: EdgeInsets.only(left: Insets.x2_5 * index),
-        decoration: BoxDecoration(
-          color: fromHex(color),
-          shape: BoxShape.circle,
-        ))
-  ]);
-}
 
-Color fromHex(String hexString) {
-  final buffer = StringBuffer();
-  if (hexString.length == 6 || hexString.length == 7) buffer.write('ff');
-  buffer.write(hexString.replaceFirst('#', ''));
-  return Color(int.parse(buffer.toString(), radix: 16));
-}
